backend/requirements.txt
asgiref==3.6.0
Django==4.1.6
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
mysql==0.0.3
mysqlclient==2.1.1
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
